# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://127.0.0.1:8080","description":"Generated server url"}],"paths":{"/api/users/{id}":{"get":{"tags":["system-controller"],"operationId":"user","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiUser"}}}}}},"put":{"tags":["system-controller"],"operationId":"updateUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserInput"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiVoid"}}}}}},"delete":{"tags":["system-controller"],"operationId":"deleteUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiVoid"}}}}}}},"/api/users/{id}/roles":{"put":{"tags":["system-controller"],"operationId":"assign","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Roles"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiVoid"}}}}}}},"/api/roles/{id}":{"put":{"tags":["system-controller"],"operationId":"updateRole","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoleInput"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiVoid"}}}}}},"delete":{"tags":["system-controller"],"operationId":"deleteRole","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiVoid"}}}}}}},"/api/permissions/{id}":{"put":{"tags":["system-controller"],"operationId":"updatePerm","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PermissionInput"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiVoid"}}}}}},"delete":{"tags":["system-controller"],"operationId":"deletePerm","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiVoid"}}}}}}},"/api/dictionaries/{id}":{"put":{"tags":["system-controller"],"operationId":"updateDict","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DictInput"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiVoid"}}}}}},"delete":{"tags":["system-controller"],"operationId":"deleteDict","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiVoid"}}}}}}},"/api/departments/{id}":{"put":{"tags":["system-controller"],"operationId":"updateDept","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeptInput"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiVoid"}}}}}},"delete":{"tags":["system-controller"],"operationId":"deleteDept","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiVoid"}}}}}}},"/api/users":{"get":{"tags":["system-controller"],"operationId":"users","parameters":[{"name":"q","in":"query","required":false,"schema":{"type":"string","default":""}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":1}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiPageUser"}}}}}},"post":{"tags":["system-controller"],"operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserInput"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiLong"}}}}}}},"/api/roles":{"get":{"tags":["system-controller"],"operationId":"roles","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiListRole"}}}}}},"post":{"tags":["system-controller"],"operationId":"createRole","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoleInput"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiVoid"}}}}}}},"/api/permissions":{"get":{"tags":["system-controller"],"operationId":"permissions","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiListPermission"}}}}}},"post":{"tags":["system-controller"],"operationId":"createPerm","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PermissionInput"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiVoid"}}}}}}},"/api/dictionaries":{"get":{"tags":["system-controller"],"operationId":"dictionaries","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiListDict"}}}}}},"post":{"tags":["system-controller"],"operationId":"createDict","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DictInput"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiVoid"}}}}}}},"/api/departments":{"get":{"tags":["system-controller"],"operationId":"departments","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiListDept"}}}}}},"post":{"tags":["system-controller"],"operationId":"createDept","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeptInput"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiVoid"}}}}}}},"/api/auth/logout":{"post":{"tags":["auth-controller"],"operationId":"logout","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiVoid"}}}}}}},"/api/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Login"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiLoginResult"}}}}}}},"/api/account/password":{"post":{"tags":["system-controller"],"operationId":"password","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PasswordInput"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiVoid"}}}}}}},"/api/overview":{"get":{"tags":["system-controller"],"operationId":"overview","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiOverview"}}}}}}},"/api/menus":{"get":{"tags":["system-controller"],"operationId":"menus","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiListPermission"}}}}}}},"/api/dictionaries/type/{type}":{"get":{"tags":["system-controller"],"operationId":"dictType","parameters":[{"name":"type","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiListDict"}}}}}}},"/api/auth/me":{"get":{"tags":["auth-controller"],"operationId":"me","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiMe"}}}}}}},"/api/audit-logs":{"get":{"tags":["system-controller"],"operationId":"audits","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":1}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiPageAudit"}}}}}}}},"components":{"schemas":{"UserInput":{"type":"object","properties":{"username":{"type":"string","minLength":1,"pattern":"[a-zA-Z0-9_]{3,32}"},"displayName":{"type":"string","maxLength":64,"minLength":0},"deptId":{"type":"integer","format":"int64"},"enabled":{"type":"boolean"},"password":{"type":"string","maxLength":72,"minLength":0}},"required":["displayName","username"]},"ApiVoid":{"type":"object","properties":{"data":{},"message":{"type":"string"}}},"Roles":{"type":"object","properties":{"roleIds":{"type":"array","items":{"type":"integer","format":"int64"},"maxItems":20,"minItems":0}},"required":["roleIds"]},"RoleInput":{"type":"object","properties":{"code":{"type":"string","minLength":1,"pattern":"[a-z][a-z0-9_]{1,31}"},"name":{"type":"string","maxLength":64,"minLength":0},"dataScope":{"type":"string","pattern":"ALL|DEPT_TREE|DEPT|SELF"},"enabled":{"type":"boolean"},"permissionIds":{"type":"array","items":{"type":"integer","format":"int64"},"maxItems":100,"minItems":0}},"required":["code","dataScope","name","permissionIds"]},"PermissionInput":{"type":"object","properties":{"code":{"type":"string","minLength":1,"pattern":"[a-z][a-z0-9:_-]{1,100}"},"name":{"type":"string","maxLength":64,"minLength":0},"type":{"type":"string","pattern":"MENU|BUTTON|API"},"path":{"type":"string","maxLength":128,"minLength":0},"parentId":{"type":"integer","format":"int64"}},"required":["code","name","type"]},"DictInput":{"type":"object","properties":{"type":{"type":"string","minLength":1,"pattern":"[a-z][a-z0-9_]{1,63}"},"label":{"type":"string","maxLength":64,"minLength":0},"value":{"type":"string","maxLength":64,"minLength":0},"sort":{"type":"integer","format":"int32","maximum":10000,"minimum":0},"enabled":{"type":"boolean"}},"required":["label","type","value"]},"DeptInput":{"type":"object","properties":{"parentId":{"type":"integer","format":"int64"},"name":{"type":"string","maxLength":64,"minLength":0}},"required":["name"]},"ApiLong":{"type":"object","properties":{"data":{"type":"integer","format":"int64"},"message":{"type":"string"}}},"Login":{"type":"object","properties":{"username":{"type":"string","maxLength":64,"minLength":0},"password":{"type":"string","maxLength":72,"minLength":0}},"required":["password","username"]},"ApiLoginResult":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/LoginResult"},"message":{"type":"string"}}},"LoginResult":{"type":"object","properties":{"token":{"type":"string"},"expiresIn":{"type":"integer","format":"int64"}}},"PasswordInput":{"type":"object","properties":{"oldPassword":{"type":"string","maxLength":72,"minLength":0},"newPassword":{"type":"string","maxLength":72,"minLength":12}},"required":["newPassword","oldPassword"]},"ApiPageUser":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/PageUser"},"message":{"type":"string"}}},"PageUser":{"type":"object","properties":{"items":{"type":"array","items":{"$ref":"#/components/schemas/User"}},"total":{"type":"integer","format":"int64"},"page":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"}}},"User":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"username":{"type":"string"},"displayName":{"type":"string"},"deptId":{"type":"integer","format":"int64"},"deptName":{"type":"string"},"enabled":{"type":"boolean"},"roleIds":{"type":"array","items":{"type":"integer","format":"int64"}}}},"ApiUser":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/User"},"message":{"type":"string"}}},"ApiListRole":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/Role"}},"message":{"type":"string"}}},"Role":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"code":{"type":"string"},"name":{"type":"string"},"dataScope":{"type":"string"},"enabled":{"type":"boolean"},"permissionIds":{"type":"array","items":{"type":"integer","format":"int64"}}}},"ApiListPermission":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/Permission"}},"message":{"type":"string"}}},"Permission":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"code":{"type":"string"},"name":{"type":"string"},"type":{"type":"string"},"path":{"type":"string"},"parentId":{"type":"integer","format":"int64"}}},"ApiOverview":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/Overview"},"message":{"type":"string"}}},"Overview":{"type":"object","properties":{"users":{"type":"integer","format":"int64"},"departments":{"type":"integer","format":"int64"},"roles":{"type":"integer","format":"int64"},"operations":{"type":"integer","format":"int64"}}},"ApiListDict":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/Dict"}},"message":{"type":"string"}}},"Dict":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"type":{"type":"string"},"label":{"type":"string"},"value":{"type":"string"},"sort":{"type":"integer","format":"int32"},"enabled":{"type":"boolean"}}},"ApiListDept":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/Dept"}},"message":{"type":"string"}}},"Dept":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"parentId":{"type":"integer","format":"int64"},"name":{"type":"string"}}},"ApiMe":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/Me"},"message":{"type":"string"}}},"Me":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"username":{"type":"string"},"deptId":{"type":"integer","format":"int64"},"permissions":{"type":"array","items":{"type":"string"},"uniqueItems":true}}},"ApiPageAudit":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/PageAudit"},"message":{"type":"string"}}},"Audit":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"username":{"type":"string"},"action":{"type":"string"},"method":{"type":"string"},"outcome":{"type":"string"},"durationMs":{"type":"integer","format":"int64"},"createdAt":{"type":"string"}}},"PageAudit":{"type":"object","properties":{"items":{"type":"array","items":{"$ref":"#/components/schemas/Audit"}},"total":{"type":"integer","format":"int64"},"page":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"}}}}}}